Create db.models.skills
