+++ db/models/users.py
@@ -33,21 +33,7 @@ class User(Base):
     skills = relationship('Skill',secondary = "users_skills",  back_populates='users') 
     role = relationship("Role",  back_populates="users")
     tags = relationship('Tag', secondary='users_tags', back_populates='users')
-    stack = relationship("Stack", back_populates='users')
-
-
-   
-class Skill(Base):
-    __tablename__ = 'skills'
-    id = Column(Integer, primary_key = True, index = True)
-    name = Column(String)
-    user_id = Column(Integer, ForeignKey('users.id'))
-    created_at = Column(TIMESTAMP(timezone=True),
-                        nullable=False, server_default=text('now()'))
-    updated_at = Column(TIMESTAMP(timezone=True),
-                        nullable=False, server_default=text('now()'))
-
-    users = relationship('User', secondary = "users_skills", back_populates='skills')  
+    stack = relationship("Stack", back_populates='users')  
 
 
 class UserTag(Base):
